Major UI and UX changes
Moved Navbar into a single file and calling it as a component when required.

Added hook for loading when button is clicked and action is in progress. (Only for pages with single buttons)

Added background image and changed several css properties and bootstrap classes for UI/UX improvements

Deleted unwanted files/resources and added new required files/resources